another refactor
