+++ crypto_utils/README.md
@@ -2,7 +2,7 @@
 
 A library of some functions used by the `integrity` application.
 
-The library defines a `CID` structure and works around limitations
+The library defines a `Cid` structure and works around limitations
 of existing Elixir projects that do not deal with certain elliptical
 curve and hashes encoded in the multiformats system. Besides relying
 on Erlang's `:crypto` and `:ssh_file` applications, the library
