+++ crypto_utils/mix.exs
@@ -21,6 +21,7 @@ defmodule CryptoUtils.MixProject do
   # Run "mix help deps" to learn about dependencies.
   defp deps do
     [
+      {:jason, "~> 1.4"},
       {:multibase, "~> 0.0.1"},
       {:multicodec, "~> 0.0.2"},
       {:cbor, "~> 1.0.0"}
