+++ did_server/mix.exs
@@ -43,7 +43,7 @@ defmodule DidServer.MixProject do
       {:floki, ">= 0.30.0", only: :test},
       {:telemetry_metrics, "~> 0.6"},
       {:telemetry_poller, "~> 1.0"},
-      {:jason, "~> 1.2"},
+      {:jason, "~> 1.4"},
       {:plug_cowboy, "~> 2.5"},
       {:crypto_utils, path: "../crypto_utils"}
     ]
