--- integrity/config/config.exs
@@ -1,11 +1,0 @@
-# This file is responsible for configuring your application
-# and its dependencies with the aid of the Config module.
-#
-# This configuration file is loaded before any dependency and
-# is restricted to this project.
-
-# General application configuration
-import Config
-
-# Import config of the dependent application.
-import_config "../../did_server/config/config.exs"
