+++ integrity/mix.exs
@@ -3,7 +3,7 @@ defmodule Integrity.MixProject do
 
   def project do
     [
-      app: :integrity_proofs,
+      app: :integrity,
       version: "0.1.0",
       elixir: "~> 1.14",
       elixirc_paths: elixirc_paths(Mix.env()),
@@ -15,7 +15,6 @@ defmodule Integrity.MixProject do
   # Run "mix help compile.app" to learn about applications.
   def application do
     [
-      mod: {Integrity.Application, []},
       extra_applications: [:logger, :crypto, :public_key, :ssh]
     ]
   end
@@ -33,8 +32,7 @@ defmodule Integrity.MixProject do
       {:cbor, "~> 1.0.0"},
       {:jcs, git: "https://github.com/pzingg/jcs.git"},
       {:test_server, "~> 0.1", only: :test},
-      {:crypto_utils, path: "../crypto_utils"},
-      {:did_server, path: "../did_server", runtime: false}
+      {:crypto_utils, path: "../crypto_utils"}
       # {:dep_from_hexpm, "~> 0.3.0"},
       # {:dep_from_git, git: "https://github.com/elixir-lang/my_dep.git", tag: "0.1.0"}
     ]
